kamaji-api/requirements.txt
amqp==1.4.8
ansible==2.1.0.0
anyjson==0.3.3
appdirs==1.4.0
celery==3.1.19
Django==1.9.7
djangorestframework==3.4.1
djangorestframework-jwt==1.7.2
Jinja2==2.8
jsonpatch==1.12
jsonpointer==1.10
jsonschema==2.5.1
kombu==3.0.30
MarkupSafe==0.23
lettuce==0.2.23
netaddr==0.7.18
ntplib==0.3.3
os-client-config==1.12.0
PyYAML==3.11
raven==5.32.0
python-Levenshtein==0.12.0
requests==2.8.1
requestsexceptions==1.1.1
six==1.10.0
wheel==0.24.0
django-requestlogging==1.0.1
mock==1.3.0
SQLAlchemy==1.0.9
pdnsapi==0.3.0b3
python-cephlibs==0.94.5.post1
sshpubkeys==2.2.0
